add PersistentSongRecord for song caching
